@@ -3,21 +3,17 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using MVCStore.Models;
+using MVCStore.Data;
 
 namespace MVCStore.Controllers
 {
     public class ProductController : Controller
     {
         private IProductRepository repository;
+        
         public ProductController(IProductRepository repo)
         {
             repository = repo;
-        }
-
-        public IActionResult Index()
-        {
-            return View();
         }
 
         public IActionResult List()
