@@ -6,16 +6,21 @@ using System.Threading.Tasks;
 
 namespace MVCStore.Data
 {
-    public class EFProductRepository
-        : IProductRepository
-    {
-        private ApplicationDbContext context;
+	public class EFProductRepository : IProductRepository
+	{
+		private ApplicationDbContext context;
 
-        public EFProductRepository(ApplicationDbContext ctx)
-        {
-            context = ctx;
-        }
+		public EFProductRepository(ApplicationDbContext ctx)
+		{
+			context = ctx;
+		}
 
-        public IQueryable<Product> Products => context.Products;
-    }
+		public IQueryable<Product> Products
+		{
+			get
+			{
+				return context.Products;
+			}
+		}
+	}
 }
